Split hp-bridge into two releases: static only and legacy-spa
After updating to v2.1.38, I tried building a custom version of the AWS
SDK again, and this time it worked without interaction fault
with hp-bridge.

The AWS JavaScript SDK is 2x the size of a custom build
that only supports S3.  For static websites like sigal and octopress,
they do not need the full AWS SDK.  Further, saving 200 KB is generally
worth the effort on the internet.

However, there are cases where you are hosting a legacy single page
application that uses the AWS SDK. To avoid confusion on why
calls to DynamoDB, for example, aren't working we also are building
with the official Amazon AWS SDK.

So, there are now two versions.